+++ openssl-sys/src/handwritten/evp.rs
@@ -652,16 +652,3 @@ extern "C" {
     pub fn EVP_EncodeBlock(dst: *mut c_uchar, src: *const c_uchar, src_len: c_int) -> c_int;
     pub fn EVP_DecodeBlock(dst: *mut c_uchar, src: *const c_uchar, src_len: c_int) -> c_int;
 }
-
-extern "C" {
-    #[cfg(ossl300)]
-    pub fn OSSL_PARAM_construct_uint(key: *const c_char, buf: *mut c_uint) -> OSSL_PARAM;
-    #[cfg(ossl300)]
-    pub fn OSSL_PARAM_construct_utf8_string(
-        key: *const c_char,
-        buf: *mut c_char,
-        bsize: size_t,
-    ) -> OSSL_PARAM;
-    #[cfg(ossl300)]
-    pub fn OSSL_PARAM_construct_end() -> OSSL_PARAM;
-}
