+++ openssl-sys/src/handwritten/mod.rs
@@ -15,6 +15,7 @@ pub use self::hmac::*;
 pub use self::kdf::*;
 pub use self::object::*;
 pub use self::ocsp::*;
+pub use self::params::*;
 pub use self::pem::*;
 pub use self::pkcs12::*;
 pub use self::pkcs7::*;
@@ -51,6 +52,7 @@ mod hmac;
 mod kdf;
 mod object;
 mod ocsp;
+mod params;
 mod pem;
 mod pkcs12;
 mod pkcs7;
